+++ NEWS.md
@@ -5,6 +5,7 @@ CRSSIO 0.8.1.9000
 
 - reindex now fails if start year cannot be successfully converted to numeric. (#103)
 - now importing xts to ensure `reindex()`, `print.nfd()`, etc. work properly. (#107)
+- added `as.data.frame.nfd()` to create data.frames from `nfd`, `crss_nf`, and `crssi` objects.
 
 CRSSIO 0.8.1
 =================